+++ packages/ragbits-core/tests/unit/vector_stores/test_qdrant.py
@@ -6,6 +6,7 @@ from pydantic import ValidationError
 from qdrant_client.http import models
 
 from ragbits.core.embeddings.noop import NoopEmbedder
+from ragbits.core.utils.pydantic import _pydantic_bytes_to_hex
 from ragbits.core.vector_stores.base import VectorStoreEntry
 from ragbits.core.vector_stores.qdrant import QdrantVectorStore
 
@@ -13,7 +14,9 @@ from ragbits.core.vector_stores.qdrant import QdrantVectorStore
 @pytest.fixture
 def mock_qdrant_store() -> QdrantVectorStore:
     return QdrantVectorStore(
-        client=AsyncMock(), index_name="test_collection", embedder=NoopEmbedder(return_values=[[[0.1, 0.2, 0.3]]])
+        client=AsyncMock(),
+        index_name="test_collection",
+        embedder=NoopEmbedder(return_values=[[[0.1, 0.2, 0.3]]], image_return_values=[[[0.7, 0.8, 0.9]]]),
     )
 
 
@@ -30,7 +33,20 @@ async def test_store(mock_qdrant_store: QdrantVectorStore) -> None:
                     "document_type": "test_type",
                 },
             },
-        )
+        ),
+        VectorStoreEntry(
+            id=UUID("827cad0b-058f-4b85-b8ed-ac741948d502"),
+            text="some other key",
+            image_bytes=b"image",
+            metadata={
+                "content": "test content",
+                "document_meta": {
+                    "title": "test title",
+                    "source": {"path": "/test/path"},
+                    "document_type": "test_type",
+                },
+            },
+        ),
     ]
 
     mock_qdrant_store._client.collection_exists.return_value = False  # type: ignore
@@ -43,9 +59,9 @@ async def test_store(mock_qdrant_store: QdrantVectorStore) -> None:
     call_points = list(call_kwargs["points"])
 
     assert call_kwargs["collection_name"] == "test_collection"
-    assert len(call_points) == 1
+    assert len(call_points) == 3
     assert call_points[0].id == "1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"
-    assert call_points[0].vector == {"text": [0.1, 0.2, 0.3]}
+    assert call_points[0].vector == [0.1, 0.2, 0.3]
     assert call_points[0].payload == {
         "id": UUID("1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"),
         "text": "test_key",
@@ -58,6 +74,36 @@ async def test_store(mock_qdrant_store: QdrantVectorStore) -> None:
             },
         },
     }
+    assert call_points[1].id == "827cad0b-058f-4b85-b8ed-ac741948d502"
+    assert call_points[1].vector == [0.1, 0.2, 0.3]
+    assert call_points[1].payload == {
+        "id": UUID("827cad0b-058f-4b85-b8ed-ac741948d502"),
+        "text": "some other key",
+        "metadata": {
+            "content": "test content",
+            "document_meta": {
+                "title": "test title",
+                "source": {"path": "/test/path"},
+                "document_type": "test_type",
+            },
+        },
+        "image_bytes": _pydantic_bytes_to_hex(b"image"),
+    }
+    assert call_points[2].id == "827cad0b-058f-4b85-b8ed-ac741948d503"
+    assert call_points[2].vector == [0.7, 0.8, 0.9]
+    assert call_points[2].payload == {
+        "id": UUID("827cad0b-058f-4b85-b8ed-ac741948d502"),
+        "text": "some other key",
+        "metadata": {
+            "content": "test content",
+            "document_meta": {
+                "title": "test title",
+                "source": {"path": "/test/path"},
+                "document_type": "test_type",
+            },
+        },
+        "image_bytes": _pydantic_bytes_to_hex(b"image"),
+    }
 
 
 async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
@@ -66,7 +112,7 @@ async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
             models.ScoredPoint(
                 version=1,
                 id="1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
-                vector={"text": [0.12, 0.25, 0.29]},
+                vector=[0.12, 0.25, 0.29],
                 score=0.9,
                 payload={
                     "id": "1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
@@ -84,11 +130,31 @@ async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
             models.ScoredPoint(
                 version=1,
                 id="827cad0b-058f-4b85-b8ed-ac741948d502",
-                vector={"text": [0.13, 0.26, 0.30]},
-                score=0.9,
+                vector=[0.7, 0.8, 0.9],
+                score=0.7,
                 payload={
                     "id": "827cad0b-058f-4b85-b8ed-ac741948d502",
                     "text": "test_key 2",
+                    "image_bytes": _pydantic_bytes_to_hex(b"image"),
+                    "metadata": {
+                        "content": "test content 2",
+                        "document_meta": {
+                            "title": "test title 2",
+                            "source": {"path": "/test/path-2"},
+                            "document_type": "txt",
+                        },
+                    },
+                },
+            ),
+            models.ScoredPoint(
+                version=1,
+                id="827cad0b-058f-4b85-b8ed-ac741948d503",
+                vector={"text": [0.13, 0.26, 0.30]},
+                score=0.5,
+                payload={
+                    "id": "827cad0b-058f-4b85-b8ed-ac741948d502",
+                    "text": "test_key 2",
+                    "image_bytes": _pydantic_bytes_to_hex(b"image"),
                     "metadata": {
                         "content": "test content 2",
                         "document_meta": {
@@ -103,8 +169,13 @@ async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
     )
 
     results = [
-        {"content": "test content 1", "title": "test title 1", "vector": [0.12, 0.25, 0.29]},
-        {"content": "test content 2", "title": "test title 2", "vector": [0.13, 0.26, 0.30]},
+        {"content": "test content 1", "title": "test title 1", "vector": [0.12, 0.25, 0.29], "score": 0.9},
+        {
+            "content": "test content 2",
+            "title": "test title 2",
+            "vector": [0.7, 0.8, 0.9],
+            "score": 0.7,
+        },
     ]
 
     query_results = await mock_qdrant_store.retrieve("query")
@@ -113,7 +184,8 @@ async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
     for query_result, result in zip(query_results, results, strict=True):
         assert query_result.entry.metadata["content"] == result["content"]
         assert query_result.entry.metadata["document_meta"]["title"] == result["title"]
-        assert query_result.vectors["text"] == result["vector"]
+        assert query_result.vector == result["vector"]
+        assert query_result.score == result["score"]
 
 
 async def test_remove(mock_qdrant_store: QdrantVectorStore) -> None:
@@ -125,7 +197,7 @@ async def test_remove(mock_qdrant_store: QdrantVectorStore) -> None:
     mock_qdrant_store._client.delete.assert_called_with(  # type: ignore
         collection_name="test_collection",
         points_selector=models.PointIdsList(
-            points=["1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"],
+            points=["1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8", "1c7d6b27-4ef1-537c-ad7c-676edb8bc8a9"],  # text and image
         ),
     )
 
@@ -137,7 +209,7 @@ async def test_list(mock_qdrant_store: QdrantVectorStore) -> None:
             models.ScoredPoint(
                 version=1,
                 id="1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
-                vector={"text": [0.12, 0.25, 0.29]},
+                vector=[0.12, 0.25, 0.29],
                 score=0.9,
                 payload={
                     "id": "1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
@@ -155,11 +227,31 @@ async def test_list(mock_qdrant_store: QdrantVectorStore) -> None:
             models.ScoredPoint(
                 version=1,
                 id="827cad0b-058f-4b85-b8ed-ac741948d502",
-                vector={"text": [0.13, 0.26, 0.30]},
-                score=0.9,
+                vector=[0.13, 0.26, 0.30],
+                score=0.7,
                 payload={
                     "id": "827cad0b-058f-4b85-b8ed-ac741948d502",
                     "text": "test_key 2",
+                    "image_bytes": _pydantic_bytes_to_hex(b"image"),
+                    "metadata": {
+                        "content": "test content 2",
+                        "document_meta": {
+                            "title": "test title 2",
+                            "source": {"path": "/test/path-2"},
+                            "document_type": "txt",
+                        },
+                    },
+                },
+            ),
+            models.ScoredPoint(
+                version=1,
+                id="827cad0b-058f-4b85-b8ed-ac741948d503",
+                vector=[0.7, 0.8, 0.9],
+                score=0.5,
+                payload={
+                    "id": "827cad0b-058f-4b85-b8ed-ac741948d502",
+                    "text": "test_key 2",
+                    "image_bytes": _pydantic_bytes_to_hex(b"image"),
                     "metadata": {
                         "content": "test content 2",
                         "document_meta": {
@@ -173,17 +265,23 @@ async def test_list(mock_qdrant_store: QdrantVectorStore) -> None:
         ]
     )
 
-    results = [
-        {"content": "test content 1", "title": "test title 1", "vector": [0.12, 0.25, 0.29]},
-        {"content": "test content 2", "title": "test title 2", "vector": [0.13, 0.26, 0.30]},
+    results: list[dict] = [
+        {"content": "test content 1", "title": "test title 1", "image": None},
+        {
+            "content": "test content 2",
+            "title": "test title 2",
+            "image": b"image",
+        },
     ]
 
     entries = await mock_qdrant_store.list(where={"document_type": "txt"})
 
     assert len(entries) == len(results)
-    for entry, result in zip(entries, results, strict=True):
+    for i, (entry, result) in enumerate(zip(entries, results, strict=True)):
         assert entry.metadata["content"] == result["content"]
         assert entry.metadata["document_meta"]["title"] == result["title"]
+        assert entry.image_bytes == result["image"]
+        assert entry.text == f"test_key {i + 1}"
 
 
 def test_create_qdrant_filter() -> None: